pyproject.toml
[project]
name = "rad-test"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "genanki>=0.13.1",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
